Fix narrowing conversion warning in struct fstring
Warning C4267: 'initializing': conversion from 'size_t' to 'const int',
possible loss of data.

Client applications fail to compile if warnings are treated as errors.

+++ include/fmt/base.h
@@ -2634,7 +2634,7 @@ inline auto runtime(string_view s) -> runtime_format_string<> { return {{s}}; }
 /// A compile-time format string.
 template <typename... T> struct fstring {
  private:
-  static constexpr int num_static_named_args =
+  static constexpr size_t num_static_named_args =
       detail::count_static_named_args<T...>();
 
   using checker = detail::format_string_checker<
